requirements.txt
django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework-api-key==3.0.0